refactor(apigateway): split service.rs into 11 resource impl files
service.rs grew to 3413 lines after Phase I extraction. Already had
multiple impl ApiGatewayService blocks (one per resource); extract
each non-setup block to a sibling file via #[path] modules:

- service_rest_api_resources.rs (760 lines): handle_control dispatch +
  rest_api/resource/method/method_response CRUD.
- service_integrations.rs (354 lines): integration + integration_response
  + test_invoke_method/authorizer.
- service_deployments.rs (163 lines)
- service_stages.rs (182 lines)
- service_models.rs (138 lines)
- service_request_validators.rs (118 lines)
- service_authorizers.rs (164 lines)
- service_api_keys.rs (148 lines)
- service_usage_plans.rs (216 lines): usage plan CRUD + plan keys + usage.
- service_vpc_links_etc.rs (566 lines): vpc_link, base_path, domain_name.
- service_extras.rs (435 lines): doc parts/versions, gateway responses,
  client cert, sdk, export.

Methods exposed as pub(super) so inline test mod (when present) can
call them. service.rs: 3413 -> 324 lines (-91%).
